fix: address PRRT_kwDOSJAM6s6JcnXf — move GC orchestration out of the route
Extract candidate extraction, dry-run augmentation, worker-deadline
budgeting, worker-param construction, delegation, and outcome folding from
the POST /v1/service/gc route into a service-layer function
(awf.service.gc_request.run_service_gc_request). The route is now thin
request/response translation only, per the api/ no-business-logic guideline.
Logic moves 1:1, so existing route tests are repointed at the new module's
patch targets; behavior is unchanged.
